Started to add Graphics Pipeline
Made separation for Compute and Graphics.
Of course with Compute pipeline you can bind to graphics command list.
But you should not be able to bind graphics pipeline to Compute command list.
Also began to refactor code and hide the monstrosities away. Chopping in to nice and fine slices.
All because I will have to add optimizations later ... so rather chop in to pieces and hide the pieces away.
For record, still compiles and tests pass.