Implemented shader interface cache.
ShaderInterface class helps identifying unique root descriptors and optimize if we need to bind one.
Pipeline creation now can fail if root descriptor varies between shaders, no error message yet. Will crash immediately.
No heaps are set, which is the reason nothing shows on screen.
Fixing that next to new format to get bindless textures at the same time.
Needs heaps to be grouped together and a manager that knows which is which.