Compile level affects shaders and bug fixed.
Removed a deprecated test that is done already.
Added stricter hlsl checking to be always used and Release mode now really has release mode specific shader optimizations.
This could also be done dynamically depending on needs. For example, developing a shader in release mode but compiling with all the debugs enabled to see possible errors with it. Or put a specific shader in to debug mode and leave rest optimized.
Removed deprecated hlsl syntax that the strict checking found.